+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 - Fix Dogstatsd tag formatter to properly start the block with `|`.
+- Specify serializer by passing a `Serializer` instance and not a class.
 
 ## [0.1.0] - 2020-10-28